pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "delphi"
version = "0.2.0"
description = "Automated Interpretability"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["interpretability", "explainable-ai"]
dependencies = [
    "datasets",
    "orjson",
    # "sparsify@git+https://github.com/EleutherAI/sparsify",
    # "sparsify @ file:./../halutsae",
    "sparsify",
    "transformers<4.48.0", # necessary for monet
    "safetensors",
    "simple_parsing",
    "jaxtyping",
    "fire",
    "blobfile",
    "bitsandbytes",
    "flask",
    "vllm",
    "aiofiles",
    "sentence_transformers",
    "outlines==0.0.44",
    "sglang-router>=0.1.4",
    "sglang>=0.4.3.post2",
    "sgl-kernel>=0.0.3.post6",
    "flashinfer-python>=0.2.1.post2",
    "decord>=0.6.0",
    "python-multipart>=0.0.20",
    "torchao>=0.8.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pyright]
include = ["delphi*"]
reportPrivateImportUsage = false

[tool.setuptools.packages.find]
include = ["delphi*"]

[tool.uv.sources]
sparsify = { path = "../halutsae", editable=true }

[[tool.uv.index]]
url = "https://flashinfer.ai/whl/cu124/torch2.5/flashinfer-python"

[tool.ruff]
# TODO: Clean up or remove experiments folder.
exclude = ["experiments", "**/prompts/*", "**/prompts.py","**/simulator.py"]
# Avoid automatically removing unused imports in __init__.py files.
# Such imports will be flagged with a dedicated message suggesting
# that the import is either added to the module's __all__ symbol
ignore-init-module-imports = true
# Ignore F821 (undefined name) errors, as they are common in type annotations.
ignore = ["F722", "F821"]
# Same as Black.
line-length = 88
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) codes
# See https://beta.ruff.rs/docs/rules/ for more possible rules
select = ["E", "F", "I"]
